move RType to common and make api and impl independent of meta
