+++ common/src/main/scala/scalan/WrapSpec.scala
@@ -1,0 +1,4 @@
+package scalan
+
+trait WrapSpec {
+}
