+++ core/src/main/scala/scalan/dynamic/Specializations.scala
@@ -2,8 +2,7 @@ package scalan.dynamic
 
 import java.lang.reflect.Method
 
-import scalan.meta.TypeDesc
-import scalan.{Lazy, _}
+import scalan.{Lazy, TypeDesc, _}
 import scalan.universe.api.TypesApi
 import scalan.universe.api.UniverseUtils._
 import scalan.util.CollectionUtil._
