+++ core/src/main/scala/scalan/impl/ViewsImpl.scala
@@ -1,9 +1,9 @@
 package scalan
 
 import java.lang.reflect.Method
+
 import scala.language.higherKinds
-import scala.collection.mutable.{Map=>MutMap}
-import scalan.meta.TypeDesc
+import scala.collection.mutable.{Map => MutMap}
 import scala.reflect.runtime.universe.{WeakTypeTag, weakTypeTag}
 import scalan.meta.ScalanAst._
