+++ core/src/main/scala/scalan/primitives/Structs.scala
@@ -1,11 +1,11 @@
 package scalan.primitives
 
-import scalan._
+import scalan.{TypeDesc, _}
+
 import scala.reflect.runtime.universe._
 import scalan.util.CollectionUtil
 import OverloadHack._
 import scalan.compilation.GraphVizConfig
-import scalan.meta.TypeDesc
 
 /**
  The code is inspired by LMS structs and is used in Scalan with the same semantics
