+++ core/src/test/scala/scalan/universe/api/ElemTests.scala
@@ -1,9 +1,8 @@
 package scalan.universe.api
 
 import scala.language.reflectiveCalls
-import scalan._
+import scalan.{TypeDesc, _}
 import scalan.common.{SegmentsModule, KindsModule, CommonExamples, ViewExamples}
-import scalan.meta.TypeDesc
 import scalan.universe.api.UniverseUtils._
 
 class ElemTests extends BaseCtxTests { suite =>
