+++ library-api/src/main/scala/special/collection/Types.scala
@@ -1,6 +1,6 @@
 package special.collection
 
-import scalan.meta.RType
+import scalan.RType
 
 import scala.reflect.ClassTag
