--- meta/src/main/scala/special/wrappers/WrapSpec.scala
@@ -1,4 +1,0 @@
-package special.wrappers
-
-trait WrapSpec {
-}
