+++ toolkit/src/main/scala/scalan/json/ScalanJsonContext.scala
@@ -5,9 +5,9 @@ import java.lang.reflect.Method
 import spray.json.{JsObject, JsArray, JsString, JsValue, JsBoolean}
 
 import scala.collection.Seq
-import scalan.Scalan
+import scalan.{Scalan, TypeDesc}
+
 import scala.collection.mutable.{Map => MMap}
-import scalan.meta.TypeDesc
 
 trait ScalanJsonContext[C <: Scalan] { self: ScalanJsonProtocol[C] =>
   val ctx: C
